system_code/backend/requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.2.0
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.2.0
certifi==2023.11.17
charset-normalizer==3.3.2
distro==1.9.0
Django==4.1.13
django-cors-headers==4.3.0
djangorestframework==3.14.0
djongo==1.3.6
dnspython==2.5.0
exceptiongroup==1.2.0
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
idna==3.6
multidict==6.0.4
openai==1.10.0
Pillow==10.1.0
pydantic==2.5.3
pydantic_core==2.14.6
pymongo==3.12.3
python-dotenv==1.0.0
pytz==2023.3.post1
requests==2.31.0
sniffio==1.3.0
sqlparse==0.2.4
tqdm==4.66.1
typing_extensions==4.9.0
urllib3==2.1.0
yarl==1.9.4
